refactor: update notification context import paths for consistency
